Fix intent tests
Two intent tests were incorrectly failing.  One was due to the
recent change which allowed a perfect-match Padatious intent
to win instead of Adapt, like it did previously.  So the a
different (but still correct) intent handler was being invoked.

The second change was due to it looking for multiple response
entities.  The test framework really only handles a single one
at the moment.